# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Operator360 Portal Backend API
  version: 1.0.0
  description: |
    Contract reference for the Operator360 portal backend, generated from the
    current handler implementations on branch f/oid-pid. All routes are under
    `/api` and require a bearer/session token validated by AuthMiddleware.

    Endpoints whose S3 file paths are keyed by an operator now resolve the S3
    prefix via `opt_master.data_path` looked up by `opt_id` — they no longer
    accept or require `opt_state` / `opt_district`.
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...
  responses:
    Unauthorized:
      description: User not found in context
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ServerError:
      description: Internal server error (S3, DB, or parse failure)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  parameters:
    OptId:
      name: opt_id
      in: query
      required: true
      schema:
        type: string
      example: MH_WMIT_SN_NS046496
...
paths:
...
  /operator_details:
    get:
      summary: Get an operator's opt_details.json
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        opt_state/opt_district are no longer accepted.
      operationId: GetOperatorDetails
      parameters:
        - $ref: "#/components/parameters/OptId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  regional_office: { type: string }
                  operator_id: { type: string }
                  file: { type: string }
                  data: { type: object, additionalProperties: true }
                  requested_by: { type: string }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path or file not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
  /operator_features:
    get:
      summary: Get an operator's transformed feature/KPI data
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        opt_state/opt_district are no longer accepted. Note the success
        response is NOT wrapped — the transformed object is the top-level body.
      operationId: GetOperatorFeatures
      parameters:
        - $ref: "#/components/parameters/OptId"
      responses:
        "200":
          description: OK (unwrapped body)
          content:
            application/json:
              schema:
                type: object
                properties:
                  opt_id: { description: "Type depends on source data, typically string" }
                  metadata:
                    type: object
                    additionalProperties: true
                    description: Present only if source JSON had a metadata key.
                  kpis:
                    type: array
                    items:
                      type: object
                      additionalProperties:
                        type: array
                        items: { type: object, additionalProperties: true }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path or file not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /operator_feedback_stats:
    get:
      summary: Get an operator's feedback submission history and per-category stats
      description: |
        Reads every /feedback (see the /feedback POST endpoint) submission for
        this opt_id back out of ClickHouse, where a Kafka Engine + materialized
        view flattens each feedback_submitted event into one row. Returns
        yes/no/unset counts per verdict category plus the full submission
        history, for the operator profile page's feedback stats card.
      operationId: GetOperatorFeedbackStats
      parameters:
        - $ref: "#/components/parameters/OptId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  opt_id: { type: string }
                  total_feedback_count: { type: integer }
                  categories:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        yes: { type: integer }
                        no: { type: integer }
                        unset: { type: integer }
                  latest: { type: object, nullable: true, additionalProperties: true }
                  history:
                    type: array
                    items: { type: object, additionalProperties: true }
        "400":
          description: opt_id missing
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /feedback:
    post:
      summary: Submit feedback for an operator
      description: |
        File path is resolved via opt_master.data_path looked up by opt_id.
        Body is a freeform object — only opt_id is required; every field
        sent is persisted as-is to S3 (opt_state/opt_district are no longer
        required but may still be sent and will be stored).
      operationId: SubmitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opt_id]
              properties:
                opt_id: { type: string }
              additionalProperties: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  regional_office: { type: string }
                  operator_id: { type: string }
                  file_path: { type: string }
                  date: { type: string }
                  submitted_by: { type: string }
        "400":
          description: Invalid JSON or missing opt_id
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Operator data path not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...